Licensed Hardware under CC-BY-SA 4.0
When applying for Open Hardware certification I learned that the MIT
License is not well suited for hardware projects.  Keep the MIT
license for the software parts of the haxophone (which lives in a
different repository, https://github.com/jcard0na/haxo-rs) and use here
Creative Commons Share-Alike licenxe for the hardware.