Add external pull-up resistors to COL lines
The Raspberry Pi has 50k-60k internal pull-up resistors on GPIOs.  While this
is working correctly on a clean board, those values are a bit weak in the
presence of leakage paths which may appear in the presence of dirt and
humidity.  Increase the immunity to PCB leakage paths by adding lower
resistance (10k) external pull-up resistors.